Encode to temp file in case something goes wrong
Previously, if the app crashed the encoded files would be left in an
incomplete state. Then, on the next run it would see these incomplete
encoded files as 'complete' and not continue or restart the encoding.

Now, it encodes to a temp file, then renames after encoding completes.

So the encoded files are always in a 'complete' state and if the app
crashes, the temp files are simply deleted and the encoding can start
again from the beginning.