Replace deletion logic with archiving
So instead of deleting the originals, it archives them by renaming with
the suffix " - Archived".

If somebody wants to delete the originals they can just delete the files
with that suffix.

And if they want to re-encode they can remove the suffix and run again.

Reason for doing this: lossless encoding creates larger files and would
prefer to delete those after video editing and retain the original
smaller file.