Run the encoding in parallel
Idea here is not to run handbrake in parallel (it wont since I just
added a lock to it) but to run the other operations in parallel - e.g.
archiving.

If archiving to another disk (slow) then it can run handbrake on the next file
while archiving the current one.