Back to Lock + Latch design
Hmm sticking with this.

Separates constructs which is nice.

Alternative would be splitting the VideoEncoder method, but don't want
to do that... callers should have access to a complete method. Leaky
abstraction otherwise.

Lock doesn't need to be fair actually, because only a single latch is
released, there is at most 1 thread waiting on the lock at one time:

Example flow:
[Thread 1] l.a -> v.a -> l.r -> v.r
[Thread 2] l.a ->        v.a -> l.r -> v.r
[Thread 3] l.a ->               v.a -> l.r -> v.r
[Thread 4] l.a ->                      v.a -> l.r -> v.r